requirements.txt
typer==0.3.2
mysql-connector-python==2.1.0
cryptography==36.0.1
